open usbViewModel + client feature usbViewModel
