+++ android-core/src/main/res/values/strings.xml
@@ -18,4 +18,7 @@
     <string name="close_bottom_sheet" translatable="false">Close bottom sheet</string>
     <string name="turn_on_notifications" translatable="false">Turn on notifications</string>
     <string name="ask_notification_desc" translatable="false">Don\'t miss important messages like bundle upload progress and application activity</string>
+    <string name="to_settings">To Settings</string>
+    <string name="to_settings_text">To enable our app to manage app-related files on this device and your USB storage, please toggle \\\"Yes\\\" for BundleClient in your settings. We assure you that only files created by our app will be accessed. Protecting user privacy and maintaining ethical practices are our top priorities.</string>
+    <string name="reload">Reload</string>
 </resources>
